Set C binding for unix_system_job in its own C file
